requirements.txt
fastapi
uvicorn
pillow 
tensorflow==2.16.2